Area of occupancy

Compute areas of occupancy (AOO) for multiple taxa in square kilometers

AOO.computing(
  XY,
  cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  export_shp = FALSE,
  proj_type = "cea"
)

Arguments

XY

data.frame see Details.

cell_size_AOO

numeric, by default is 2. Value indicating the grid size in kilometres used for estimating Area of Occupancy.

nbe.rep.rast.AOO

numeric, by default is 0. Indicate the number of raster with random starting position used for estimating the AOO. If 0 but some translation of the raster are still done.

parallel

a logical. Whether running should be performed in parallel. FALSE by default.

NbeCores

an integer. Register the number of cores for parallel execution. Two by default.

show_progress

logical. Whether progress informations should displayed. TRUE by default

export_shp

logical, by default is FALSE. Whether a shapefile of occupied cells should be exported.

proj_type

string or numeric

Value

If export_shp if FALSE (the default) a numeric vector of AOO estimates for each taxa If export_shp if TRUE a list with two elements

a vector of AOO estimates for each taxa

a list of simple feature for each taxa

Input data

XY as a data.frame should have the following structure:

It is mandatory to respect field positions, but column names do not matter

The first column is contains numeric value i.e. latitude in decimal degrees

The second column is contains numeric value i.e. longitude in decimal degrees

The third column is contains character value i.e. the names of the species

See Examples.

Iteration to get the minimal AOO

The argument nbe.rep.rast.AOO should ideally be higher than 10 for increasing the chance to get the minimal number of occupied cell. However, increasing nbe.rep.rast.AOO also increases the computing time. Note that if nbe.rep.rast.AOO = 0, several translations of the grid overlaying occurrences are still conducted

proj_type

See proj_type

References

Gaston & Fuller 2009 The sizes of species' geographic ranges, Journal of Applied Ecology, 49 1-9

Author

Gilles Dauby gilles.dauby@ird.fr

Examples

data(dataset.ex)

AOO <- AOO.computing(dataset.ex)
#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%


# This would estimate AOO for all taxa by overlaying randomly a 
# grid 3 times. For each taxa, the minimum value is kept

AOO <- AOO.computing(dataset.ex, nbe.rep.rast.AO = 3)
#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%