Compute Extent of Occurrence

Compute EOO given a data frame of coordinate in decimals degrees

EOO.comp(
  XY,
  method.range = "convex.hull",
  alpha = 1,
  buff.alpha = 0.1,
  method.less.than3 = "not comp",
  mode = "spheroid",
  proj_type = NULL,
  reproject = FALSE
)

Arguments

XY

data.frame

method.range

string, by default "convex.hull", can also take "alpha.hull"

alpha

integer

buff.alpha

numeric

method.less.than3

string

mode

character string either 'spheroid' or 'planar'. By default 'spheroid'

proj_type

crs

reproject

logical FALSE whether the polygon should be converted to geographic coordinates if mode is planar

exclude.area

logical. Default if FALSE

country_map

SpatialPolygonDataframe. Default if NULL

Value

A list

EOO a numeric vector of AOO estimates for each taxa

spatial.polygon a simple feature collection

Author

Gilles Dauby & Renato A. Ferreira de Lima