Categorize taxa according to IUCN criterion A

Provide the consensus IUCN category based on the sub-criteria of IUCN criterion A (A1, A2, A3 and A4) and the thresholds recommended by IUCN.

cat_criterion_a(
  A1_val = NULL,
  A2_val = NULL,
  A3_val = NULL,
  A4_val = NULL,
  A1.threshold = c(50, 70, 90),
  A234.threshold = c(30, 50, 80),
  all.cats = TRUE
)

Arguments

A1_val

a numeric vector of the estimates of population decline based on IUCN sub-criterion A1

A2_val

a numeric vector of the estimates of population decline based on IUCN sub-criterion A2

A3_val

a numeric vector of the estimates of population decline based on IUCN sub-criterion A3

A4_val

a numeric vector of the estimates of population decline based on IUCN sub-criterion A4

A1.threshold

a numeric vector of threshold values for sub-criterion A1, default is c(50, 70, 90) following IUCN guidelines

A234.threshold

a numeric vector of A2, A3 and A4 threshold values for sub-criteria A2, A3 and A4, default is c(30, 50, 80) following IUCN guidelines

all.cats

a logical value indicating whether to return all categories or only the consensus, default is TRUE

Value

A list of two or three elements:

ranks_A: a character vector with the consensus category of each taxon based on subcriteria A1-A4

cats_code: a character vector with the final IUCN category for each taxon

all_cats: a data frame with the categorization of each taxon based on all sub criteria, only returned if all.cats=TRUE

Details

By default, the function provides the consensus category, following the recommendations of IUCN (2019) that states "Only the criteria for the highest category of threat that the taxon qualifies for should be listed". Therefore, the consensus category is the highest category of threat among the sub-criteria evaluated.

The function assumes that the order of the values in A1_val, A2_val, A3_val and A4_val are from the same taxa (i.e. first element from A1_val until A4_val is always the same species i). Therefore, the order of the estimates of population decline for each sub-criterion must be the same.

References

IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Author

Gilles Dauby & Renato A. Ferreira de Lima

Examples

A1_val <- c(40, 90, 75, 10)
A2_val <- c(90, 95, 30, 10)
A3_val <- c(10, 30, 45, 15)
A4_val <- c(32, 18, 80, 20)

## All four criteria
cat_criterion_a(
  A1_val = A1_val,
  A2_val = A2_val,
  A3_val = A3_val,
  A4_val = A4_val
)
#> $ranks_A
#> [1] "CR"       "CR"       "CR"       "LC or NT"
#> 
#> $cats_code
#> [1] "A2"          "A1+A2"       "A4"          "A1+A2+A3+A4"
#> 
#> $all_cats
#>         A1       A2       A3       A4
#> 1 LC or NT       CR LC or NT       VU
#> 2       CR       CR       VU LC or NT
#> 3       EN       VU       VU       CR
#> 4 LC or NT LC or NT LC or NT LC or NT
#> 

## All four criteria, no categories for all criteria, just the consesus
cat_criterion_a(
  A1_val = A1_val,
  A2_val = A2_val,
  A3_val = A3_val,
  A4_val = A4_val,
  all.cats = FALSE
)
#> $ranks_A
#> [1] "CR"       "CR"       "CR"       "LC or NT"
#> 
#> $cats_code
#> [1] "A2"          "A1+A2"       "A4"          "A1+A2+A3+A4"
#> 

## One or more criteria not evaluated
cat_criterion_a(
  A1_val = NULL,
  A2_val = A2_val,
  A3_val = A3_val,
  A4_val = A4_val
)
#> Warning: The following subcriteria were not used in the assessment:  A1
#> $ranks_A
#> [1] "CR"       "CR"       "CR"       "LC or NT"
#> 
#> $cats_code
#> [1] "A2"       "A2"       "A4"       "A2+A3+A4"
#> 
#> $all_cats
#>         A2       A3       A4
#> 1       CR LC or NT       VU
#> 2       CR       VU LC or NT
#> 3       VU       VU       CR
#> 4 LC or NT LC or NT LC or NT
#> 

cat_criterion_a(
  A1_val = A1_val,
  A2_val = NULL,
  A3_val = NULL,
  A4_val = NULL
)
#> Warning: The following subcriteria were not used in the assessment:  A2, A3, A4
#> $ranks_A
#> [1] "LC or NT" "CR"       "EN"       "LC or NT"
#> 
#> $cats_code
#> [1] "A1" "A1" "A1" "A1"
#> 

## One or more criteria not evaluated, one species
cat_criterion_a(
  A1_val = NULL,
  A2_val = 0.4,
  A3_val = NULL,
  A4_val = NULL
)
#> Warning: The following subcriteria were not used in the assessment:  A1, A3, A4
#> $ranks_A
#> [1] "LC or NT"
#> 
#> $cats_code
#> [1] "A2"
#>