Categorize taxa according to IUCN criterion B

Provide IUCN threat categories based on B sub-criteria, conditions and thresholds.

cat_criterion_b(
  EOO = NULL,
  AOO = NULL,
  locations = NULL,
  sever.frag = NULL,
  protected = NULL,
  decline = NULL,
  ext.fluct = NULL,
  EOO.threshold = c(20000, 5000, 100),
  AOO.threshold = c(2000, 500, 10),
  Loc.threshold = c(10, 5, 1),
  protected.threshold = 100,
  fluct.threshold = 10,
  all.cats = TRUE
)

Arguments

EOO

numeric vector with species extent of occurrence - EOO (i.e. sub-criterion B1)

AOO

numeric vector with species area of occupancy - AOO (i.e. sub-criterion B2)

locations

numeric vector with the number of locations where the species occur (i.e. condition 'a' of criterion B)

sever.frag

logical. Is the population severely fragmented? If TRUE the condition 'a' of criterion B will be met.

protected

numeric vector providing estimated percentage of species distribution in protected areas

decline

string vector providing the status of the species continuing decline in EOO, AOO, habitat, locations or subpopulations or population size (i.e. condition 'b'). If different of 'Decreasing', the condition 'b' of criterion B will not be met.

ext.fluct

numeric vector with the mean order of magnitude of the differences between population minima and maxima (currently not implemented).

EOO.threshold

numeric vector with the EOO thresholds to convert estimates into threat categories. Default is the threshold recommended by IUCN.

AOO.threshold

numeric vector with the AOO thresholds to convert estimates into threat categories. Default is the thresholds recommended by IUCN.

Loc.threshold

numeric vector with the thresholds of number of locations (condition 'a'). Default is the thresholds recommended by IUCN.

protected.threshold

numeric, one value indicating the threshold for protected value above which a taxa would not be threatened whatever the others parameters, by default is 100

fluct.threshold

numeric. Threshold of the order of magnitude of the differences between population minima and maxima to classify extreme fluctuations. Default to 10 as recommended by IUCN.

all.cats

logical. Should the categories from all criteria be returned and not just the consensus categories? Default to TRUE.

Value

A list of three or four elements:

ranks_B: a character vector with the consensus category of each taxon based on subcriteria B1 and B2

cats_code: a character vector with the final IUCN category for each taxon

orig_ranks_B: original vector of threat categories for each taxon before taking into account the threshold for protected areas

all_cats: a data frame with the categorization of each taxon based on all sub criteria, only returned if all.cats=TRUE

Details

The function categorizes taxa following criterion B and categories of the IUCN.

References

IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Author

Gilles Dauby & Renato A. Ferreira de Lima

Examples

EOO <- c(34000, 5000)
AOO <- c(300, 25)
locations <- c(9, 12)
cat_criterion_b(EOO = EOO,AOO = AOO, locations = locations)
#> No information on continuing decline is provided: assumed to be true
#> $ranks_B
#> [1] "VU"       "LC or NT"
#> 
#> $cats_code
#> [1] "B2a" NA   
#> 
#> $all_cats
#>        B1a      B2a
#> 1 LC or NT       VU
#> 2 LC or NT LC or NT
#> 

EOO <- c(34000, 80)
AOO <- c(300, 25)
locations <- c(1, 1)
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations)
#> No information on continuing decline is provided: assumed to be true
#> $ranks_B
#> [1] "EN" "CR"
#> 
#> $cats_code
#> [1] "B2a" "B1a"
#> 
#> $all_cats
#>        B1a B2a
#> 1 LC or NT  EN
#> 2       CR  EN
#> 

EOO <- c(50, 5000)
AOO <- c(5, 25)
locations <- c(1, 10)
protected <- c(80, 50)
decline <- c("Stable", "Decreasing")
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations, 
                  protected = protected, decline = decline)
#> Some taxa categorized as threatened based on EOO/AOO/locations were finally assessed as not threatened because no continuing decline was detected
#> $ranks_B
#> [1] "LC or NT" "VU"      
#> 
#> $cats_code
#> [1] NA     "B2ab"
#> 
#> $all_cats
#>   B1a B2a
#> 1  CR  CR
#> 2  VU  VU
#> 

EOO <- c(34000, 5000)
AOO <- c(300, 25)
locations <- c(9, 12)
sever.frag <- c(FALSE, TRUE)
protected <- c(100, 80)
decline <- c("Decreasing", "Decreasing")
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations, 
                  sever.frag = sever.frag, protected = protected, decline = decline)
#> Some taxa categorized as Threatened finally assessed as Not Threatened because percentage of their area in protected areas above the protected.threshold
#> $ranks_B
#> [1] "LC or NT" "EN"      
#> 
#> $cats_code
#> [1] NA     "B2ab"
#> 
#> $orig_ranks_B
#> [1] "VU" "EN"
#> 
#> $all_cats
#>        B1a B2a
#> 1 LC or NT  VU
#> 2       VU  EN
#> 

EOO <- c(34000, 5000)
AOO <- c(300, 25)
locations <- c(15, 12)
sever.frag <- c(TRUE, TRUE)
protected <- c(80, 50)
decline <- c("Decreasing", "Decreasing")
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations, 
                  sever.frag = sever.frag, protected = protected, decline = decline)
#> $ranks_B
#> [1] "EN" "EN"
#> 
#> $cats_code
#> [1] "B2ab" "B2ab"
#> 
#> $all_cats
#>        B1a B2a
#> 1 LC or NT  EN
#> 2       VU  EN
#>