Downlist Threat Categories

Perform the downlisting of threat categories which is often necessary for regional conservation assessments (IUCN 2012).

cat_downlist(cats = NULL, down.by = 1, dd = FALSE, sign = TRUE)

Arguments

cats

character. The vector containing the IUCN threat categories.

down.by

numerical. The number of steps to downlist the categories. Default to 1.

dd

logical. Should the Data Deficient (DD) category be included in the downlisting? Default to FALSE.

sign

logical. Should the degrees sign be indicated in the downlisted category. Default to TRUE.

Details

For regional conservation assessments, the IUCN recommends an additional step, which relates to the possible "effect of populations of the same taxon in neighbouring regions on the regional population", and thus to the possibility of a possible rescue effect (IUCN 2012).

Although the IUCN (2012) considers the exceptional possibility of uplisting, the function only performs downlisting of the categories of threat. Downlisting normally is a one-step change in the category (i.e. from EN to VU), but this can be controlled by the argument down.by (default to 1).

By default, the Data Deficient category is excluded from the downlisting. The Least Concern category remains unaltered as well.

Note that "if it is unknown whether or not extra-regional populations influence the extinction risk of the regional population, the category (...) should be kept unaltered" (IUCN 2012).

References

IUCN (2012). Guidelines for Application of IUCN Red List Criteria at Regional and National Levels (Version 4.0). IUCN. Gland, Switzerland and Cambridge, UK. 41pp.

Author

Renato A. Ferreira de Lima

Examples

cats <- c("EX","CR","EN","VU","NT","DD","LC","NA","NE")
cat_downlist(cats)
#> [1] "CRo" "ENo" "VUo" "NTo" "LCo" "DD"  "LCo" "NA"  "NE" 
cat_downlist(cats, down.by = 2)
#> [1] "ENo" "VUo" "NTo" "LCo" "LCo" "DD"  "LCo" "NA"  "NE" 
cat_downlist(cats, dd = TRUE)
#> [1] "CRo" "ENo" "VUo" "NTo" "LCo" "LCo" "LCo" "NA"  "NE" 
cat_downlist(cats, sign = FALSE)
#> [1] "CR" "EN" "VU" "NT" "LC" "DD" "LC" "NA" "NE"
cat_downlist(cats, down.by = 2, dd = TRUE, sign = FALSE)
#> [1] "EN" "VU" "NT" "LC" "LC" "LC" "LC" "NA" "NE"