Categorize Taxa Using Multiple IUCN Criteria

Provide the consensus IUCN category based on multiples IUCN sub-criteria.

cat_mult_criteria(assess.df = NULL, evidence.df = NULL)

Arguments

assess.df

a data frame containing the taxon name in the first columns and the assessments under each criterion in the subsequent columns.

evidence.df

a data frame (not currently implemented)

Value

The same data frame as assess.df with three new columns: the consensus category ('category'), the main criteria that lead to this category ('main.criteria'). It also returns the auxiliary category provided by other criteria ('aux.criteria'), separated by a ';'

Details

The definition of the main category of threat, follows the recommendations of IUCN (2019) that states "Only the criteria for the highest category of threat that the taxon qualifies for should be listed". Therefore, the consensus category is the highest category of threat among the sub-criteria evaluated. Nevertheless, the function also returns the categories and sub-criteria related to lower categories of threat.

References

IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Author

Renato A. Ferreira de Lima

Examples

df <- data.frame(tax = c("sp1","sp2","sp3","sp4"),
                    A2 = c(NA, "VU", "VU", NA),
                    B1 = c("LC", "VU", "LC", "LC"),
                    B2 = c("LC", "EN", "LC", "VU"),
                    D = c(NA, "LC", "LC", NA))
cat_mult_criteria(df)                     
#>   tax   A2 B1 B2    D category main.criteria aux.criteria
#> 1 sp1 <NA> LC LC <NA>       LC         B1+B2             
#> 2 sp2   VU VU EN   LC       EN            B2    VU: A2+B1
#> 3 sp3   VU LC LC   LC       VU            A2    LC: B1+B2
#> 4 sp4 <NA> LC VU <NA>       VU            B2       LC: B1