Assess IUCN Criterion B

Preliminary assessment of species conservation status following IUCN Criterion B, which is based on species geographic distribution (i.e. extent of occurrence - EOO, and area of occupancy, AOO)

criterion_B(
  x = NULL,
  AOO = NULL,
  EOO = NULL,
  locations = NULL,
  severe.frag = NULL,
  subpops = NULL,
  decline = NULL,
  EOO.threshold = c(20000, 5000, 100),
  AOO.threshold = c(2000, 500, 10),
  Loc.threshold = c(10, 5, 1),
  comp_subpop = TRUE,
  comp_severe.frag = FALSE,
  resol_sub_pop = 5,
  cell_size_locations = 10,
  method_locations = "fixed_grid",
  method_polygons = "no_more_than_one",
  rel_cell_size = 0.05,
  threat_list = NULL,
  names_threat = NULL,
  threat_weight = NULL,
  id_shape = "id_orig",
  country_map = NULL,
  method.range = "convex.hull",
  alpha = 1,
  buff.alpha = 0.1,
  exclude.area = FALSE,
  cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  threshold_severe = 50,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2,
  proj_type = "cea",
  mode = "spheroid",
  DrawMap = FALSE,
  add.legend = TRUE
)

Arguments

x

a dataframe or an object of class spgeoIN see https://github.com/azizka/speciesgeocodeR. See Details

AOO

a vector of species AOO, if available

EOO

a vector of species EOO, if available

locations

a vector of species number of locations, if available

severe.frag

a vector indicating if species is severely fragmented, if available

subpops

a vector of the number of subpopulations (sensu IUCN), if available

decline

a vector providing the status of the species continuing decline in EOO, AOO, habitat, locations or subpopulations or population size (i.e. condition 'b') to be passed on to function cat_criterion_b(). If different of 'Decreasing', the condition 'b' of criterion B will not be met.

EOO.threshold

numeric vector indicating the thresholds used to categorize EOO in IUCN categories

AOO.threshold

numeric vector indicating the thresholds used to categorize AOO in IUCN categories

Loc.threshold

numeric vector indicating the thresholds used to categorize the number of locations in IUCN categories

comp_subpop

logical if the number of sub-populations should be computed. By default is TRUE

comp_severe.frag

logical if severe fragmentation should be computed. By default is FALSE

resol_sub_pop

a value defining the radius of the circles around each occurrence (in kilometres) or data frame vector containing a column 'tax' with the taxa names and a column 'radius' with the species-specific radius (in kilometre as well). Typically, this data frame is the output of ConR function subpop.radius.

cell_size_locations

numeric, value indicating the grid size in kilometres used for estimating the number of location. By default, equal to 10

method_locations

string, indicating the method used for estimating the number of locations. See Details

"fixed_grid" (by default)

"sliding_scale"

method_polygons

string. Used if threat_list is provided. See Details.

"no_more_than_one" (the default): each single POLYGON will be considered as a single location

"grid": a grid of cell_size_locations size will be used to estimate the number of location within polygons

rel_cell_size

numeric, if method_locations="sliding_scale", cell_size_locations is ignored and the resolution is given by the maximum distance separating two occurrences multiplied by rel_cell_size. By default, it is 0.05

threat_list

list or sfc objects POLYGON or MULTIPOLYGON documenting threats. If provided, this will be taken into account for calculating number of location (see Details and method_polygons). By default, no shapefile is provided

names_threat

character vector, indicating names of threats, optional

threat_weight

numeric vector, indicating weight given to each threat

id_shape

string

country_map

a SpatialPolygonsDataFrame or SpatialPolygons showing for example countries or continent borders. This shapefile will be used for cropping the SpatialPolygonsl if exclude.area is TRUE

method.range

a character string, "convex.hull" or "alpha.hull". By default is "convex.hull"

alpha

a numeric, if method.range is "alpha.hull", value of alpha of the alpha hull, see alphahull::ahull(). By default is 1

buff.alpha

a numeric, if method.range is "alpha.hull", define the buffer in decimal degree added to alpha hull. By default is 0.1

exclude.area

a logical, if TRUE, areas outside of country_map are cropped of SpatialPolygons used for calculating EOO. By default is FALSE

cell_size_AOO

numeric, by default is 2. Value indicating the grid size in kilometres used for estimating Area of Occupancy.

nbe.rep.rast.AOO

numeric, by default is 0. Indicate the number of raster with random starting position used for estimating the AOO. If 0 but some translation of the raster are still done.

threshold_severe

numeric of one value indicates the threshold to identify severe fragmentation. By default is 50

parallel

logical, whether running in parallel. By default, it is FALSE

show_progress

logical, whether a bar showing progress in computation should be shown. By default, it is TRUE

NbeCores

integer, register the number of cores for parallel execution. By default, it is 2

proj_type

string or numeric

mode

character string either 'spheroid' or 'planar'. By default 'spheroid'

DrawMap

logical, by default is FALSE, if TRUE, a png map is created for each species in a directory of the working environment

add.legend

logical, whether legend should be added to map

Value

A data frame containing, for each of taxon, (EOO, AOO, n.locs, n.subpops?), the IUCN categories associated with the sub-criteria and the consensus category for criterion B.

References

IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Author

Gilles Dauby & Renato A. Ferreira de Lima