Assess IUCN Criterion C

Preliminary assessment of species conservation status following IUCN Criterion C, which is based on small population size (e.g. <10,000 mature individuals) that are declining or may decline in the near future.

criterion_C(
  x,
  years = NULL,
  assess.year = NULL,
  project.years = NULL,
  project = TRUE,
  ignore.years = NULL,
  recent.year = NULL,
  generation.time = NULL,
  prop.mature = NULL,
  subpop.size = NULL,
  models = c("linear", "quadratic", "exponential", "logistic", "general_logistic",
    "piecewise"),
  subcriteria = c("C1", "C2"),
  correction = NULL,
  C.threshold = c(10000, 2500, 250),
  C1.threshold = c(10, 20, 25),
  C2ai.threshold = c(1000, 250, 50),
  C2aii.threshold = c(90, 95, 100),
  mag.fluct = 10,
  high.alter = 80,
  all.cats = TRUE,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  ...
)

Arguments

x

a vector (one species) or a data frame (multiple species/ subpopulations) containing the population sizes (e.g. number of mature individuals) per year, from the oldest to the most recent estimate.

years

a vector containing the years for which the population sizes is available (i.e. time series). Can be NULL if x contain years as names.

assess.year

the year for which the assessment should be performed.

project.years

a vector containing the years for which population sizes were or should be projected.

project

logical. Should population sizes be projected into the future? Default to TRUE.

ignore.years

any year(s) that should be ignored for calculating continuing decline of populations?

recent.year

the year to be used as a the starting year used to assess recent continuing decline (see details).

generation.time

a value or vector of generation lengths, i.e. the average age of parents of the current cohort (IUCN 2019).

prop.mature

a value or vector of the proportion of mature individuals in the total population (IUCN 2019). Default to 1.

subpop.size

a named list containing the vector of number of mature individuals per subpopulation. The length of the list must match the length and order of the taxa being assessed.

models

a vector containing the names of the models to be fitted to species population data to perform projections.

subcriteria

a vector containing the sub-criteria that should be included in the assessment (i.e. C1 and/or C2).

correction

a value or vector of correction values, that should applyed to the reduction in population size estimated from 'x'.

C.threshold

numeric vector with the criterion C thresholds to define small population sizes (e.g. number of mature individuals). Default values are the thresholds recommended by the IUCN.

C1.threshold

numeric vector with the C1 thresholds of continuing decline. Default values are the thresholds recommended by the IUCN.

C2ai.threshold

numeric vector with the C2a i thresholds for the population size of the largest subpopulation. Default are the values recommended by the IUCN.

C2aii.threshold

numeric vector with the C2a ii thresholds for the percentage of the population size in the same subpopulation. Default are the values recommended by the IUCN.

mag.fluct

numerical. Threshold of mean order of magnitude of the differences between population minima and maxima to classify populations with extreme fluctuations. Default to 10 as recommended by IUCN (2019).

high.alter

numerical. Threshold of proportion of changes that are followed by a change in the opposite direction. Default to 80%. Currently NOT implemented.

all.cats

logical. Should the categories from all criteria be returned and not just the consensus categories?

parallel

logical. Should calculations be parallelized? Default to FALSE.

NbeCores

integer. Number of cores for parallel computing. Default to 2.

show_progress

logical. Should the progress bar be displayed? Default to TRUE.

...

other parameters to be passed as arguments for function pop.decline.fit

Value

A data frame containing, for each of taxon, the year of assessment, the time interval of the assessment (include past and future estimates, if any), the population sizes in the interval of assessment, the model used to obtain the projections of population size, the population decline and subpopulation descritors related to sub-criteria C1 and C2, the IUCN categories associated with these sub-criteria and the consensus category for criterion C.

Details

The function criterion_C is similar to another ConR function: criterion_A. The main difference between these functions relies on the differences between criteria A and C as described by IUCN (2019, p.70): "criterion C applies only to small populations, the time frame over which the decline is measured is shorter (...) and the decline rate thresholds are lower, because the populations are already small".

Two basic tests are performed for each taxon for the assessment of criterion C. First, we test if the population is small. By default, we use the maximum value of the thresholds recommended by IUCN (2019): 10,000 mature individuals. If the taxon is not below this threshold, the assessment is not performed. IUCN (2019) does not specify at what time the population size should be below the threshold. Here, we consider the year of the assessment.

Next, we test if population size is actually declining. IUCN (2019, p.43) defines: "A continuing decline is a recent, current or projected future decline (...) which is liable to continue unless remedial measures are taken. (...). Continuing declines at any rate can be used to qualify taxa under criteria B or C2. Estimated continuing decline (under criterion C1) has quantitative thresholds, and requires a quantitative estimate, which can be calculated using the same methods as for population reduction" (i.e. criterion A). Therefore, function criterion_C consider two types of decline: (i) continuing decline at any rate (sub-criterion C2) and (ii) estimated continuing decline (sub-criterion C1).

The first type of decline is defined based on the mean change of population size between observations (no statistical fit); if the mean change from the first population size suggests a decline in the population size, then the population is classified as declining. The user have to provide the year to delimit the period considered to be recent, which should include threatening processes that are representative or indicative of present-day patterns. Although IUCN (2019) considers declines at any rate, here we consider populations in decline those with an average decline of 0.1% or more, in order to incorporate small fluctuations in stable populations. Moreover, although (IUCN 2019, p.43) states that under criteria C2, "continuing declines can be observed, estimated, inferred or projected", here we consider only observed, estimated, inferred before the years of assessment.

The second type of decline is defined on the statistical models fitted to the observed and/or projected population data. Once the best model is selected, the confidence interval of the parameters is computed. If the parameter estimates indicate a declining trend, then the population is classified as declining (e.g. the slope parameter of the linear model is negative, as well as the confidence interval around the slope estimate). For this type of decline, we consider observed, estimated or projected (IUCN 2019).

In the case of taxa with population size per subpopulation, there are two ways to entering subpopulation information. The first is to provide a named list with a vector of population sizes of each species at the year of assessment. The other is to provide population sizes for each subpopulation in x, and repeat the name of the taxon in the first column of x. In the case of subpopulations, the overall reduction in population size is obtained as recommended by IUCN (2019, p.38) which is average reduction across all subpopulation, weighted by their initial sizes.

As defined by IUCN (2019, p. 44), extreme fluctuations are variations in population size or area typically greater than one order of magnitude. In addition, "Fluctuations must be inferred only where there is reasonable certainty that a population change will be followed by a change in the reverse direction within a generation or two" IUCN (2019).

The argument prop.mature can be used if the population data provided are not already the number of mature individuals (i.e. population size sensu IUCN, 2019). By default, the proportion of mature individuals in the total population proportion is taken as 1, but the user can provide one proportion for all species or species- specific proportions.

The argument correction applies any correction desired to the reduction obtained from the vector(s) of population sizes per year provided in x for 1, 2 and 3 generation times, related to the assessment of sub-criterion C1 (the correction currently does not apply to the input population size vector and consequently to the population size at the time of assessment or the maximum size of subpopulations - see the help of function criterion_A() for an example on when one should apply this correction). Here, values should be positive and if one or more species do not need for correction just enter the value one. Values between zero and one will reduced the value of population size reduction and values above one will increase them.

References

IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Author

Renato A. Ferreira de Lima & Gilles Dauby

Examples

  
  ## Example with subpopulations
  data(example_criterionC_subpops)
  
  criterion_C(x = example_criterionC_subpops,
  years = NULL, 
  assess.year = 2000,
  project.years = NULL,
  generation.time = 10,
  subpop.size = NULL,
  models = c("linear", "exponential", "logistic", "general_logistic"),
  subcriteria = c("C1", "C2")
  )
#> Warning: The years of the population sizes were not given and were taken from the input population data
#> Warning: Starting year to define recent continuing decline not in the provided. Assuming the middle of the time series: 1985
#> Warning: Only one generation length provided for two or more taxa: assuming the same generation length for all taxa
#> Computing the estimated continuing decline (subcriteria C1)...
#>   |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  11%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |================                                                      |  22%  |                                                                              |=======================                                               |  33%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |===============================                                       |  44%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |=======================================                               |  56%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================                |  78%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |==============================================================        |  89%  |                                                                              |======================================================================| 100%
#>         tax assessment.year assessment.period assessment.pop.sizes
#> 1 species 1            2000    1970-2000-2030          248-227-206
#> 2 species 2            2000    1970-2000-2030          855-555-333
#> 3 species 3            2000    1970-2000-2030            135-49-18
#> 4 species 4            2000    1970-2000-2030       9200-8000-7024
#> 5 species 5            2000    1970-2000-2030     22500-11000-5058
#> 6 species 6            2000    1970-2000-2030   22227600-1245000-0
#> 7 species 7            2000    1970-2000-2030          210-229-252
#> 8 species 8            2000    1970-2000-2030     23000-13900-4960
#> 9 species 9            2000    1970-2000-2030            150-97-15
#>   predictive.model assess.pop.size any.decline cont.decline reduction_3gen
#> 1      exponential             227  Decreasing   Decreasing       8.467742
#> 2      exponential             555  Decreasing   Decreasing      35.087719
#> 3      exponential              49  Decreasing   Decreasing      63.703704
#> 4      exponential            8000  Decreasing   Decreasing      13.043478
#> 5      exponential           11000  Decreasing   Decreasing      51.111111
#> 6         logistic         1245000  Decreasing   Decreasing      94.398855
#> 7      exponential             229  Increasing   Increasing      -9.047619
#> 8           linear           13900  Decreasing   Decreasing      39.565217
#> 9           linear              97  Increasing       Stable      35.333333
#>   reduction_2gen reduction_1gen max.subpop.size prop.subpop.size
#> 1       5.020921       1.731602             100         44.05286
#> 2      24.489796       6.722689             500         90.09009
#> 3      54.205607      27.941176              45         91.83673
#> 4       8.571429       5.325444            6000         75.00000
#> 5      36.416185      18.518519            9000         81.81818
#> 6      79.844585      41.521841         1200000         96.38554
#> 7      -4.566210      -1.777778             100         43.66812
#> 8      30.845771      19.186047            7300         52.51799
#> 9      34.013605      19.834711              82         84.53608
#>   mean.fluctuation alternance       C1       C2 category_C category_C_code
#> 1             0.01         50 LC or NT       EN         EN            C2ai
#> 2             0.08         50       EN       VU         EN              C1
#> 3             0.19         50       CR       CR         CR   C1+C2ai+C2aii
#> 4             0.02         50       VU LC or NT         VU              C1
#> 5             0.13         50 LC or NT LC or NT   LC or NT                
#> 6             0.72         50 LC or NT LC or NT   LC or NT                
#> 7             0.01         50 LC or NT LC or NT   LC or NT                
#> 8             0.09         50 LC or NT LC or NT   LC or NT                
#> 9            10.56        100 LC or NT LC or NT   LC or NT                
  
  ## Same example, but using the argument `prop.mature` 
  criterion_C(x = example_criterionC_subpops,
  years = NULL, 
  assess.year = 2000,
  project.years = NULL,
  generation.time = 10,
  prop.mature = 0.85,
  subpop.size = NULL,
  models = c("linear", "exponential", "logistic", "general_logistic"),
  subcriteria = c("C1", "C2")
  )
#> Warning: The years of the population sizes were not given and were taken from the input population data
#> Warning: Starting year to define recent continuing decline not in the provided. Assuming the middle of the time series: 1985
#> Warning: Only one generation length provided for two or more taxa: assuming the same generation length for all taxa
#> Warning: Only one proportion of mature individuals provided for two or more taxa: assuming the same proportion for all taxa
#> Computing the estimated continuing decline (subcriteria C1)...
#>   |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  11%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |================                                                      |  22%  |                                                                              |=======================                                               |  33%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |===============================                                       |  44%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |=======================================                               |  56%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================                |  78%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |==============================================================        |  89%  |                                                                              |======================================================================| 100%
#>         tax assessment.year assessment.period assessment.pop.sizes
#> 1 species 1            2000    1970-2000-2030          211-193-175
#> 2 species 2            2000    1970-2000-2030          727-472-283
#> 3 species 3            2000    1970-2000-2030            115-42-15
#> 4 species 4            2000    1970-2000-2030       7820-6800-5970
#> 5 species 5            2000    1970-2000-2030      19125-9350-4299
#> 6 species 6            2000    1970-2000-2030   18893460-1058250-0
#> 7 species 7            2000    1970-2000-2030          178-195-214
#> 8 species 8            2000    1970-2000-2030     19550-11815-4216
#> 9 species 9            2000    1970-2000-2030            128-82-12
#>   predictive.model assess.pop.size any.decline cont.decline reduction_3gen
#> 1      exponential          192.95  Decreasing   Decreasing       8.467742
#> 2      exponential          471.75  Decreasing   Decreasing      35.087719
#> 3      exponential           41.65  Decreasing   Decreasing      63.703704
#> 4      exponential         6800.00  Decreasing   Decreasing      13.043478
#> 5      exponential         9350.00  Decreasing   Decreasing      51.111111
#> 6         logistic      1058250.00  Decreasing   Decreasing      94.398855
#> 7      exponential          194.65  Increasing   Increasing      -9.047619
#> 8           linear        11815.00  Decreasing   Decreasing      39.565217
#> 9           linear           82.45  Increasing       Stable      35.333333
#>   reduction_2gen reduction_1gen max.subpop.size prop.subpop.size
#> 1       5.020921       1.731602             100         44.05286
#> 2      24.489796       6.722689             500         90.09009
#> 3      54.205607      27.941176              45         91.83673
#> 4       8.571429       5.325444            6000         75.00000
#> 5      36.416185      18.518519            9000         81.81818
#> 6      79.844585      41.521841         1200000         96.38554
#> 7      -4.566210      -1.777778             100         43.66812
#> 8      30.845771      19.186047            7300         52.51799
#> 9      34.013605      19.834711              82         84.53608
#>   mean.fluctuation alternance       C1       C2 category_C category_C_code
#> 1             0.01         50 LC or NT       EN         EN            C2ai
#> 2             0.08         50       EN       VU         EN              C1
#> 3             0.19         50       CR       CR         CR   C1+C2ai+C2aii
#> 4             0.02         50       VU LC or NT         VU              C1
#> 5             0.13         50       VU LC or NT         VU              C1
#> 6             0.72         50 LC or NT LC or NT   LC or NT                
#> 7             0.01         60 LC or NT LC or NT   LC or NT                
#> 8             0.09         50 LC or NT LC or NT   LC or NT                
#> 9            10.56        100 LC or NT LC or NT   LC or NT                
  
  ## Example without subpopulations (cannot assess subcriteria C2)
  data(example_criterionC)
  
  criterion_C(x = example_criterionC,
  years = NULL, 
  assess.year = 2000,
  project.years = NULL,
  generation.time = 10,
  subpop.size = NULL,
  models = c("linear", "quadratic", "exponential", "logistic", "general_logistic"),
  subcriteria = c("C1")
  )
#> Warning: The years of the population sizes were not given and were taken from the input population data
#> Warning: Starting year to define recent continuing decline not in the provided. Assuming the middle of the time series: 1985
#> Warning: Only one generation length provided for two or more taxa: assuming the same generation length for all taxa
#> Computing the estimated continuing decline (subcriteria C1)...
#>   |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  11%  |                                                                              |================                                                      |  22%  |                                                                              |=======================                                               |  33%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |===============================                                       |  44%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#>   |                                                                              |=======================================                               |  56%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================                |  78%  |                                                                              |==============================================================        |  89%  |                                                                              |======================================================================| 100%
#> Warning: A linear or quadratic model was also selected among the candidate models but was not considered as the best model
#> 
#> Warning: The following subcriteria were not used in the assessment:  C2
#>        tax assessment.year assessment.period assessment.pop.sizes
#> 1 species1            2000    1970-2000-2030              248-227
#> 2 species2            2000    1970-2000-2030          855-555-333
#> 3 species3            2000    1970-2000-2030            135-49-18
#> 4 species4            2000    1970-2000-2030       9200-8000-7024
#> 5 species5            2000    1970-2000-2030     22500-11000-5058
#> 6 species6            2000    1970-2000-2030   22227600-1245000-0
#> 7 species7            2000    1970-2000-2030          210-229-227
#> 8 species8            2000    1970-2000-2030     23000-13900-4960
#> 9 species9            2000    1970-2000-2030            150-97-28
#>   predictive.model assess.pop.size any.decline cont.decline reduction_3gen
#> 1        quadratic             227  Decreasing   Decreasing       8.467742
#> 2      exponential             555  Decreasing   Decreasing      35.087719
#> 3      exponential              49  Decreasing   Decreasing      63.703704
#> 4      exponential            8000  Decreasing   Decreasing      13.043478
#> 5      exponential           11000  Decreasing   Decreasing      51.111111
#> 6         logistic         1245000  Decreasing   Decreasing      94.398855
#> 7        quadratic             229  Increasing   Increasing      -9.047619
#> 8           linear           13900  Decreasing   Decreasing      39.565217
#> 9      exponential              97  Increasing       Stable      35.333333
#>   reduction_2gen reduction_1gen       C1 category_C category_C_code
#> 1       5.020921       1.731602 LC or NT   LC or NT                
#> 2      24.489796       6.722689       EN         EN              C1
#> 3      54.205607      27.941176       CR         CR              C1
#> 4       8.571429       5.325444       VU         VU              C1
#> 5      36.416185      18.518519 LC or NT   LC or NT                
#> 6      79.844585      41.521841 LC or NT   LC or NT                
#> 7      -4.566210      -1.777778 LC or NT   LC or NT                
#> 8      30.845771      19.186047 LC or NT   LC or NT                
#> 9      34.013605      19.834711 LC or NT   LC or NT