Assess IUCN Criterion D

Preliminary assessment of species conservation status following IUCN Criterion D, which is based on very small or restricted populations (Subcriteria D1 and D2).

criterion_D(
  pop.size = NULL,
  name_sp = NULL,
  AOO = NULL,
  n.Locs = NULL,
  prop.mature = NULL,
  subcriteria = c("D", "D2"),
  D.threshold = c(1000, 250, 50),
  AOO.threshold = NULL,
  Loc.threshold = NULL,
  all.cats = TRUE
)

Arguments

pop.size

a value (one species) or a vector (multiple species/ subpopulations) containing the population sizes (e.g. number of mature individuals).

name_sp

a vector containing the name of the taxa to be assessed. Default to "Species 1", "Species 2", ..., "Species n", where n is the number of taxa.

AOO

a vector containing the Area of Occupancy of each taxon (in km2).

n.Locs

a vector containing the number of locations for each taxon.

prop.mature

a value or vector of the proportion of mature individuals in the total population (IUCN 2019). Default to 1.

subcriteria

a vector containing the sub-criteria that should be included in the assessment (i.e. D and/or D2).

D.threshold

numeric vector with the criterion D thresholds of very small population sizes. Default values are the thresholds recommended by the IUCN.

AOO.threshold

numeric vector containing the threshold of the Area of Occupancy (AOO) in km2. Default to NULL.

Loc.threshold

numeric vector containing the threshold of the Number of Locations. Default to NULL.

all.cats

logical. Should the categories from all criteria be returned and not just the consensus categories?

Value

A data frame containing the name of each taxon, the population sizes, the AOO and Number of Locations, the IUCN categories associated with each sub-criterion and the consensus category for criterion D.

Details

This is a simple and fast function to assess IUCN criterion D. The assessment based solely on population sizes is done automatically, if a vector of population sizes is provided. Caution should be taken while assessing sub-criterion D2. IUCN (2019, p.71) emphasizes that there is no strict thresholds for D2 and that this sub-criterion should only be assessed if the "population is prone to the effects of human activities or stochastic events in an uncertain future, and is thus capable of becoming Critically Endangered or even Extinct in a very short time period. (...). So, simply meeting the suggested (or any other) threshold for AOO or number of locations is not sufficient". Therefore, criterion_D assumes no default thresholds for D2.

The argument prop.mature can be used if the population data provided are not already the number of mature individuals (i.e. population size sensu IUCN, 2019). By default, the proportion of mature individuals in the total population proportion is taken as 1, but the user can provide one proportion for all species or species- specific proportions.

Currently, the function does not supports data separated by subpopulation.

References

IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Author

Renato A. Ferreira de Lima & Gilles Dauby

Examples

## All sub-criteria
criterion_D(pop.size = c(25, 50, 200, 250, 500, 1000, 1500), 
            AOO = c(1,2,5,10,15,20,25), 
            n.Locs = c(1,1,2,3,4,5,6),
            subcriteria = c("D", "D2"),
            AOO.threshold = 20,
            Loc.threshold = 5)
#>         tax pop.size AOO Numb.Locations        D   D2.Loc    D.AOO category_D
#> 1 Species 1       25   1              1       CR       VU       VU         CR
#> 2 Species 2       50   2              1       EN       VU       VU         EN
#> 3 Species 3      200   5              2       EN       VU       VU         EN
#> 4 Species 4      250  10              3       VU       VU       VU         VU
#> 5 Species 5      500  15              4       VU       VU       VU         VU
#> 6 Species 6     1000  20              5 LC or NT LC or NT LC or NT   LC or NT
#> 7 Species 7     1500  25              6 LC or NT LC or NT LC or NT   LC or NT
#>   category_D_code
#> 1               D
#> 2               D
#> 3               D
#> 4  D+D2.Loc+D.AOO
#> 5  D+D2.Loc+D.AOO
#> 6                
#> 7                
            
## Only sub-criterion D
criterion_D(pop.size = c(25, 50, 200, 250, 500, 1000, 1500), 
            subcriteria = c("D"))
#>         tax pop.size        D category_D category_D_code
#> 1 Species 1       25       CR         CR               D
#> 2 Species 2       50       EN         EN               D
#> 3 Species 3      200       EN         EN               D
#> 4 Species 4      250       VU         VU               D
#> 5 Species 5      500       VU         VU               D
#> 6 Species 6     1000 LC or NT   LC or NT                
#> 7 Species 7     1500 LC or NT   LC or NT                
            
## Olny sub-criterion D2 (Please read the function details before its use)            
criterion_D(pop.size = NULL, 
            AOO = c(1,2,5,10,15,20,25), 
            n.Locs = c(1,1,2,3,4,5,6),
            subcriteria = c("D2"),
            AOO.threshold = 20,
            Loc.threshold = 5)
#>         tax AOO Numb.Locations   D2.Loc    D.AOO category_D category_D_code
#> 1 Species 1   1              1       VU       VU         VU    D2.Loc+D.AOO
#> 2 Species 2   2              1       VU       VU         VU    D2.Loc+D.AOO
#> 3 Species 3   5              2       VU       VU         VU    D2.Loc+D.AOO
#> 4 Species 4  10              3       VU       VU         VU    D2.Loc+D.AOO
#> 5 Species 5  15              4       VU       VU         VU    D2.Loc+D.AOO
#> 6 Species 6  20              5 LC or NT LC or NT   LC or NT                
#> 7 Species 7  25              6 LC or NT LC or NT   LC or NT