Test Estimated Continuing Decline

Based on statistical models fitted to the population size data, this function assess if the model parameter estimates suggest a continuing decline.

pop.decline.test(x, best.name = NULL, assess.year = NULL)

Arguments

x

the object containing the model fitted to population data. Tipically the result from ConR function pop.decline.fit

best.name

name of the the model fitted to data. Not used if x is the result from function pop.decline.fit

assess.year

the year for which the assessment should be performed

Details

The function extracts the confidence interval of the parameters of the model selected to describe the trend in population size. If the confidence interval of the parameters suggests a significant decline, the trend is classified as 'significantly decreasing'. If there is evidence of a decline, but the evidence is not significant, the trend is classified as 'non-significantly decreasing'. For instance, if the inclination parameter of the linear model is negative and its confidence interval does not include zero, the trend is 'significantly decreasing'; if the inclination is negative and the confidence interval include zero, the trend is 'non-significantly decreasing'.

For the particular case of the piecewise-regression model, the function returns the classification of the population trend for each time interval.

If the number of observations is too small (n. obs. <7) to obtain confidence intervals for models with more than two parameters (e.g. genealized logistic model), the assessment is carried out empirically, by assessing if the model predictions provides values that are sucessively declining. In this case, the trend is just classified as 'increasing' or 'decreasing', and thus test of 'estimated continuing decline' becomes the same as the test of 'continuing decline at any rate' (sub-criterion B2).

All significance tests of population trends assume a confidence level of 0.95 (the default of stats function confint()). In the particular case of singular gradients of model fit, the function progressively decreases the confidence level from 0.95 until 0.75 until it gets estimates lower and upper confidence intervals.

References

IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Author

Renato A. Ferreira de Lima

Examples

## Creating vectors with the population data and time intervals 
#(adapted from the IUCN 2019 workbook for Criterion A, available 
#at: https://www.iucnredlist.org/resources/criterion-a)
pop = c(10000, 9600, 9100, 8200, 7500, 7200, 7000)
yrs = c(1970, 1973, 1975, 1980, 1985, 1987, 1990)

## Fitting data with different models and setting
best.model = pop.decline(pop.size = pop, years = yrs, 
 models = c("linear","exponential","logistic"), by.taxon = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
pop.decline.test(x = best.model, assess.year = 1990)
#> [1] "signif.decline"

best.model = pop.decline(pop.size = pop, years = yrs, 
 models = c("general_logistic"), by.taxon = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
pop.decline.test(x = best.model, assess.year = 1990)
#> [1] "signif.decline"

best.model = pop.decline(pop.size = pop, years = yrs, 
 models = c("quadratic"), by.taxon = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
pop.decline.test(x = best.model, assess.year = 1990)
#> [1] "signif.decline"

best.model = pop.decline(pop.size = pop, years = yrs, 
 models = c("piecewise"), by.taxon = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
pop.decline.test(x = best.model, assess.year = 1990)
#> [1] "signif.decline (1970-1980)|signif.decline (1985-1990)"