Assessment of fragmentation subpopulations

Assess whether species populations are severely fragmented

severe_frag(
  XY = NULL,
  cell_size = 2,
  nbe_rep = 0,
  AOO_poly = NULL,
  resol_sub_pop = NULL,
  habitat_poly = NULL,
  dist_isolated = NULL,
  proj_type = "cea",
  export_shp = FALSE,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2,
  threshold_severe = 50
)

Arguments

XY

data frame

cell_size

numeric

nbe_rep

numeric, by default is 0. Indicate the number of raster with random starting position used for estimating the AOO. If 0 but some translation of the raster are still done.

AOO_poly

numeric. AOO of the species in squared kilometres.

resol_sub_pop

numeric. Defines in kilometres the radius of the circles around each occurrence

habitat_poly

Simple feature collection documenting habitat of species, see Details

dist_isolated

numeric vector providing a value for each species. Distance in kilometres for identifying subpopulations considered to be isolated. See details

proj_type

string or numeric

export_shp

logical

parallel

a logical. Whether running should be performed in parallel. FALSE by default.

show_progress

logical. Whether progress informations should displayed. TRUE by default

NbeCores

an integer. Register the number of cores for parallel execution. Two by default.

threshold_severe

numeric of one value indicates the threshold to identify severe fragmentation. By default is 50

Details

This function evaluates the proportion of the total area of occupancy in habitat patches separated from others by a large distance. Based on IUCN guidelines page 48: "A taxon can be considered to be severely fragmented if most (>50%) of its total area of occupancy is in habitat patches that are (...) separated from other habitat patches by a large distance."

This function interpret subpopulations as obtained using subpop.estimation function as habitat patches. First, subpopulations that are isolated i.e. distant to all others subpopulations by at least dist_isolated kilometres, are identified. Then, the percentage of the AOO concerned by those "isolated" subpopulations is calculated.

Examples

mydf <- data.frame(ddlat = c(-44.6,-46.2,-45.4,-42.2,-43.7,-45.0,-48.0),
                   ddlon = c(-42.2,-42.6,-45.3,-42.5,-42.3,-39.0,-37.2),
                   tax = rep("a", 7),
                   stringsAsFactors = FALSE)

severe_frag(XY = mydf, dist_isolated = 200, resol_sub_pop = 10)
#> Area of occupancy computation
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Subpopulations computation
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Isolated patches computation
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%  |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   tax     frac severe_frag
#> 1   a 42.85714       FALSE