Estimate the number of subpopulations for one or multiple species

Estimate the number of subpopulations following the method circular buffer method (overlapping buffered circles form a single subpopulation)

subpop.comp(
  XY,
  resol_sub_pop = NULL,
  proj_type = "cea",
  export_shp = FALSE,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2
)

Arguments

XY

a data frame containing the geographical coordinates for each taxon (see Details).

resol_sub_pop

a value defining the radius of the circles around each occurrence (in kilometres) or data frame vector containing a column 'tax' with the taxa names and a column 'radius' with the species-specific radius (in kilometre as well). Typically, this data frame is the output of ConR function subpop.radius.

proj_type

string or numeric

export_shp

logical. Whether the resulting shapefiles should be exported. FALSE by default.

parallel

a logical. Whether running should be performed in parallel. FALSE by default.

show_progress

logical. Whether progress informations should displayed. TRUE by default

NbeCores

an integer. Register the number of cores for parallel execution. Two by default.

Value

If export_shp is TRUE,

number_subpop a numeric vector of AOO estimates for each taxa

poly_subpop a simple feature collection

a Simple feature collection with as many MULTIPOLYGON as taxa. If export_shp is FALSE, a vector with estimated number of subpopulation per taxa.

Details

XY as a data.frame should have the following structure:

It is mandatory to respect field positions, but field names do not matter

The first column is contains numeric value i.e. latitude in decimal degrees

The second column is contains numeric value i.e. longitude in decimal degrees

The third column is contains character value i.e. the names of the species

References

Rivers MC, Bachman SP, Meagher TR, Lughadha EN, Brummitt NA (2010) Subpopulations, locations and fragmentation: applying IUCN red list criteria to herbarium specimen data. Biodiversity and Conservation 19: 2071-2085. doi: 10.1007/s10531-010-9826-9

Author

Gilles Dauby & Renato A. Ferreira de Lima

Examples

data(dataset.ex)

subpop.comp(dataset.ex, resol_sub_pop = 5)
#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%
#>                      tax subpop
#> 1     Berlinia bruneelii     88
#> 2     Oncocalamus mannii     29
#> 3 Platycoryne guingangae     34
#> 4      Psychotria minuta     10
#> 5              species_1      2
#> 6              species_2      3
rad.df <- data.frame(
    tax = unique(dataset.ex$tax),
    radius = seq(3,13, by=2),
    stringsAsFactors = FALSE
  )
subpop.comp(dataset.ex, resol_sub_pop = rad.df)
#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%
#>                      tax subpop
#> 1     Berlinia bruneelii     80
#> 2     Oncocalamus mannii     29
#> 3 Platycoryne guingangae     32
#> 4      Psychotria minuta     10
#> 5              species_1      2
#> 6              species_2      2
subpop.comp(dataset.ex, resol_sub_pop = rad.df, export_shp = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%
#> $number_subpop
#>                      tax subpop
#> 1     Berlinia bruneelii     80
#> 2     Oncocalamus mannii     29
#> 3 Platycoryne guingangae     32
#> 4      Psychotria minuta     10
#> 5              species_1      2
#> 6              species_2      2
#> 
#> $poly_subpop
#> Simple feature collection with 155 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 5.655084 ymin: -15.23224 xmax: 33.79618 ymax: 5.962126
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>                   tax                       geometry
#> 1  Berlinia bruneelii POLYGON ((11.49359 -4.49189...
#> 2  Berlinia bruneelii POLYGON ((12.39218 -3.51255...
#> 3  Berlinia bruneelii POLYGON ((12.62548 -3.29588...
#> 4  Berlinia bruneelii POLYGON ((12.57689 -3.02518...
#> 5  Berlinia bruneelii POLYGON ((11.54218 -1.97918...
#> 6  Berlinia bruneelii POLYGON ((10.02548 -1.62918...
#> 7  Berlinia bruneelii POLYGON ((12.41019 -1.02515...
#> 8  Berlinia bruneelii POLYGON ((12.77548 -0.86250...
#> 9  Berlinia bruneelii POLYGON ((12.62048 -0.21116...
#> 10 Berlinia bruneelii POLYGON ((12.02548 -0.11250...
#>