Area of occupancy

Compute areas of occupancy (AOO) for multiple taxa in square kilometers

AOO.computing(
  XY,
  Cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  export_shp = FALSE
)

Arguments

XY
"dataframe" see Details

Cell_size_AOO
numeric, value indicating the grid size in kilometers used for estimating Area of Occupancy. By default, equal to 2

nbe.rep.rast.AOO
numeric , indicate the number of raster with random starting position for estimating the AOO. By default, it is 0 but some minimal translation of the raster are still done

parallel
logical, whether running in parallel. By default, it is FALSE

NbeCores
string integer, register the number of cores for parallel execution. By default, it is 2

show_progress
logical, whether a bar showing progress in computation should be shown. By default, it is TRUE

export_shp
logical, whether a shapefile of occupied cells should be exported. By default, it is FALSE

Value

If export_shp if FALSE a vector of AOO estimates for each taxa If export_shp if TRUE a list with two elements

a vector of AOO estimates for each taxa

a list of SpatialPolygonsDataFrame for each taxa

Details

Input as a dataframe should have the following structure:

It is mandatory to respect field positions, but field names do not matter

[,1]	ddlat	numeric, latitude (in decimal degrees)

[,2]	ddlon	numeric, longitude (in decimal degrees)

[,3]	tax	character or factor, taxa names

The argument of nbe.rep.rast.AOO ideally should be higher than 20 for increasing the chance to get the minimal number of occupied cell. Increasing nbe.rep.rast.AOO however also increase the computing time. So this is a trade-off that depend on the importance to get the minimal AOO and the sie of the dataset.

References

Gaston & Fuller 2009 The sizes of species'geographic ranges, Journal of Applied Ecology, 49 1-9

Examples

data(dataset.ex)
if (FALSE) {
AOO <- AOO.computing(dataset.ex)
}

# This would estimate AOO for all taxa by overlaying randomly a 
# grid 100 times. For each taxa, the minimum value is kept
if (FALSE) {
AOO <- AOO.computing(dataset.ex, nbe.rep.rast.AO = 100)
}